Added: From given 3 Float Values find the maximum

+++ TestMaximum.java
@@ -1,6 +1,7 @@
 /*
  * The Generics program implements Test cases that simply
  * finds Maximum of 3 variables using Generics.
+ * UC2 - From given 3 Float Values find the maximum.
  *
  * @author: Sanjana Bhandure
  * @version: 1.1
@@ -12,14 +13,14 @@ public class TestMaximum {
     int x, y, z;
 
     // constructor
-    public TestMaximum(int x, int y, int z) {
+    public TestMaximum(Float x, Float y, Float z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
     /*
-     * testMaximum method created for checking max of three integers
+     * testMaximum method created for checking max of three float values
      */
     public static Float testMaximum(Float x, Float y, Float z) {
         // Initializing x as greater
